backend/blog_generator/requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
